requirements.txt
fastapi==0.115.6
pandas==2.2.3
uvicorn==0.34.0
pinecone==5.4.2
PyPDF2==3.0.1
python-dotenv==1.0.1
langchain==0.3.13
huggingface-hub==0.27.0
langchain-community==0.3.13
langchain-huggingface==0.1.2
langchain-pinecone==0.2.1
python-multipart==0.0.20
